+++ apps/backoffice-v2/src/pages/Entity/hooks/useTasks/useTasks.tsx
@@ -240,6 +240,9 @@ export const useTasks = ({
                               { type, format, pattern, isEditable = true, dropdownOptions, value },
                             ]) => {
                               const fieldValue = value || (properties?.[title] ?? '');
+                              // if (value === 'bank_statement') {
+                              console.log({ props: propertiesSchema?.properties });
+                              // }
                               const isEditableDecision = isDoneWithRevision || !decision?.status;
 
                               return {
@@ -382,5 +385,7 @@ export const useTasks = ({
     pluginsOutput,
     pluginsOutputKeys,
     results,
+    noAction,
+    mutateRevisionTaskById,
   ]);
 };